requirements.txt
streamlit
ultralytics
Pillow
scikit-learn
matplotlib
torch
opencv-python-headless
